fix(test): make the published-URL gate unable to report a green it has not earned
Review follow-up. The URL fix itself is correct; the gate written to keep the
defect from returning could pass without checking anything, and lived where
none of the harness rules could see it.

The skip
--------
It read a checked-out `dist/` and `describe.skip`ped itself when that was
absent. Measured on this branch: with `dist/sitemap.xml` moved aside, the file
reported `4 skipped` and the run went green. `dist/` is gitignored, is
rewritten by `build`, `build:only` and Playwright's `VITE_E2E=1` webServer
alike, and does not exist on a fresh clone — where the documented quickstart
is `build:only`, which writes no sitemap at all. So the gate reported success
on precisely the machine where nobody had checked. Reading whichever build ran
last is only marginally better: a stale `dist/` can satisfy it while the
current tree is broken.

That it passed in CI at all was an accident of step order — the web job runs
`npm run build` before `npm run test`. Nothing in the config expressed that
dependency.

The blind spot
--------------
It also sat in `web/scripts/__tests__/`, and `allTestFiles()` in
harness-architecture.test.ts only walks `src/`. So the rule that a
process-spawning test must be declared, and the rule that the unit/build split
may not become a way to drop a file, could not see it — nor could tsconfig's
include. This is the same directory #171's follow-up (cc814798) moved the
dev-asset gate OUT of, for the same reasons, four commits before this branch
was written. Concurrent work, not a missed convention, but main has the
pattern now.

The fix
-------
Same shape as the sibling gate: it spawns its own `vite build` plus prerender
pass into a throwaway directory, is declared in PRODUCTION_BUILD_TESTS, and
lives under `src/` so the harness polices that declaration itself. It never
looks at `dist/`, so there is no artifact whose absence it could interpret as
success.

`prerender.ts` gained a `PRERENDER_OUT` override for this. Its output path was
hardcoded, so the gate would otherwise have had to clobber the developer's own
`dist/` to get a prerendered tree to assert against.

Its "dist/index.html is missing" error now names the directory it actually
looked in, which with the override is not `dist/`.

The three assertions are unchanged, and the reason for checking on disk rather
than over HTTP still holds: `vite preview` serves both the slashed and the
unslashed form, so an HTTP test would pass on the very defect that reached
production.

Verified by mutation, in both directions
----------------------------------------
Dropping the PRODUCTION_BUILD_TESTS entry:

  × no test spawns a process without being declared a build test
    → add these to PRODUCTION_BUILD_TESTS in vite.config.ts

which is the protection the move under `src/` actually bought — in
`scripts/__tests__/` that rule matched nothing.

Reverting the trailing slash in `publicHref`:

  × each page declares itself as canonical, at the address it is served from
    → canonical disagrees with the served URL:
        declares https://stabileo.com/en
        declares https://stabileo.com/en/blog
        ... (all nineteen)

which is the original defect, caught end to end through a real build.

Cost
----
85 s for the gate in the full run below — one `vite build` plus a prerender
driving headless Chromium over nineteen pages — with the build pass as a whole
at 86.5 s. Read that as an upper bound: the same machine ran the neighbouring
e2e-hook-gating in 36 s on a quiet run and 71 s in this one, so roughly half of
it is load rather than the new work. The build pass is serial for exactly this
kind of thing.

It does mean `npm test` now needs Chromium, which the web CI job already
installs before it builds.

Not addressed here
------------------
`prerender.ts` still rebuilds the URL rule by string concatenation rather than
calling `publicHref`, so the sitemap and the pages agree by review rather than
by construction. This gate is what catches a divergence, which is an argument
for having it, not for leaving the duplication. `public-routes.ts` imports
from `store.svelte.ts`, so sharing the builder with a Node script is more than
a one-line change.

Gates: `test` both passes green, 23 tests in the build pass · `typecheck` at baseline 479, with the moved file
now inside tsconfig's include · harness-architecture 11/11.
