+++ web/vite.config.ts
@@ -54,6 +54,11 @@ const PRODUCTION_BUILD_TESTS = [
   // NOTE: no apostrophes or quotes in these comments — `configArray` in
   // harness-architecture.test.ts harvests every quoted run inside this array.
   'src/lib/utils/__tests__/no-dev-assets-in-build.test.ts',
+  // Spawns a vite build AND a prerender pass, then checks that every URL the sitemap
+  // publishes is a file the host serves rather than an address it redirects. It builds
+  // its own for the same reason as the one above: reading a checked-out dist/ let it
+  // skip itself into a green on any machine where nobody had built.
+  'src/lib/i18n/__tests__/published-urls-resolve.test.ts',
 ];
